fix: Tag account & OAuth config with correct visibility
The Config interface previously used `@deepvisibility secret` for both
PagerDutyOAuthConfig & PagerDutyAccountConfig, which leads to many of
the common strings being accidently redacted in the logs.

Of particular pain is `region`, which is commonly `us`, causing every
mention of those characters, e.g. in `user` to be redacted as `***er`.

Fixing this requires declaring explicit visibility for the fields in the
type, as well as updating the fields in the Config interface.

Signed-off-by: Robin H. Johnson <rjohnson@coreweave.com>
Reference: https://github.com/PagerDuty/backstage-plugin/issues/158